FaceRec improved
