+++ src/main/java/com/packagename/prototype1/views/CameraTestView.java
@@ -1,18 +1,28 @@
 package com.packagename.prototype1.views;
 
 import com.packagename.prototype1.VideoComponent;
+import com.packagename.prototype1.backend.facerec.FaceRecognizerModule;
+import com.vaadin.flow.component.Text;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.html.Image;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.page.PendingJavaScriptResult;
+import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.router.Route;
+import org.bytedeco.javacpp.BytePointer;
+
+import javax.xml.bind.DatatypeConverter;
+import java.io.File;
+import java.io.IOException;
 
 @Route("camera")
 public class CameraTestView extends VerticalLayout {
     private Button cameraButton = new Button("Access Camera");
     private Button snapButton = new Button("Snap");
     private VideoComponent vComponent = new VideoComponent();
+
+    FaceRecognizerModule fr = new FaceRecognizerModule();
     public CameraTestView() {
         /* 
          * JavaScript Snippets:
@@ -41,7 +51,9 @@ public class CameraTestView extends VerticalLayout {
                 vComponent.getElement().getChild(0),
                 vComponent.getElement().getChild(1)
             );
+            fr.train();
         });
+
 
         snapButton.addClickListener(c -> {
             PendingJavaScriptResult res = UI.getCurrent().getPage().executeJs(
@@ -51,7 +63,17 @@ public class CameraTestView extends VerticalLayout {
             );
             res.then(String.class, dataURL -> {
                 Image snappedImage = new Image(dataURL, "snapshot");
-                add(snappedImage);
+                //add (snappedImage);
+                String str = dataURL;
+                String level = null;
+                try {
+                    level = fr.compareimage (str);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+                System.out.println(level);
+                Text tf = new Text(level);
+                add (tf);
             });
         });
     }
